# uploads/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SpinalCore Graph API",
    "version": "1.0.0",
    "description": "Welcome to the reference documentation for the Spinalcore Digital Twin REST API! </br></br>To learn about the common use cases and concept of Spinalcore REST APIs, check out our resource center  (https://resourcecenter.fr.spinalcom.com/ressources-d%C3%A9veloppeur).</br></br>In addition to Spinalcore API Reference, we also provide complete documentation for integrator that need to install and implement Spinalcore Digital Twin Operating System on their built environment (https://resourcecenter.fr.spinalcom.com/ressources-int%C3%A9grateur)",
    "termsOfService": "",
    "contact": {
      "email": "developers@spinalcom.com"
    },
    "x-logo": {
      "url": "/logo.png"
    },
    "x-preferred": true,
    "x-providerName": "spinalcom.com",
    "x-unofficialSpec": true
  },
  "components": {
    "securitySchemes": {
      "OauthSecurity": {
        "type": "oauth2",
        "description": "This API uses OAuth 2 with the implicit grant flow. [More info](https://api.example.com/docs/auth)",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "http://localhost:8080/oauth2/token",
            "scopes": {
              "read": "read request",
              "admin": "admin request"
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/context/{id}/nodeTypeList": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Return node type list of context",
        "summary": "Get type list in context with given ID context",
        "tags": ["Contexts/ontologies"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ContextNodeTypeList"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/context/{contextId}/node/{nodeId}/nodesOfType/{type}": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "return the nodes of type from a node in a context",
        "summary": "Get nodes of type from a node in a context with given IDcontext IDnode and type",
        "tags": ["Contexts/ontologies"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "nodeId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "type",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/ContextNodeofTypes"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/Network/{id}/device_list": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Return list of device",
        "summary": "Gets a list of device",
        "tags": ["IoTNetwork & Time Series"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/IoTNetwork"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/endpoint/{id}/attributsList": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "Returns list of attributs of endpoint",
        "summary": "Get list of attributs of endpoint",
        "tags": ["IoTNetwork & Time Series"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/EndPointNodeAttribut"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
    "/api/v1/endPointsGroup/{contextId}/category/{categoryId}/read": {
      "get": {
        "security": [
          {
            "OauthSecurity": ["read"]
          }
        ],
        "description": "read category endPoints Group",
        "summary": "Get category endPoints Group",
        "tags": ["EndPoints Group"],
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "path",
            "name": "categoryId",
            "description": "use the dynamic ID",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/BasicNode"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "ContextNodeofTypes": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "ContextNodeTypeList": {
      "type": "array",
      "items": {
        "type": "string"
      }
    },
...
    "IoTNetwork": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
    "EndPointNodeAttribut": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "attributs": {
          "type": "array",
          "items": {
            "anyOf": [
              {
                "$ref": "#/definitions/EndPointAttributs"
              }
            ]
          }
        }
      }
    },
    "EndPointAttributs": {
      "type": "object",
      "properties": {
        "label": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "date": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "unit": {
          "type": "integer"
        }
      }
    },
...
    "BasicNode": {
      "type": "object",
      "properties": {
        "dynamicId": {
          "type": "integer"
        },
        "staticId": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "type": {
          "type": "string"
        }
      }
    },
...
  },
...
}